# TM/Task-Management/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The API Documentation",
        "description": "This document is used to know api's status and run the api's",
        "contact": {
            "email": "arunbalajiselvan@gmail.com"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8004/public/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/source-control/store": {
            "post": {
                "tags": [
                    "Source Control"
                ],
                "summary": "Create new souce control",
                "description": "create new souce control",
                "operationId": "createSourceControl",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "githuh"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "examplelink.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/source-control/update": {
            "post": {
                "tags": [
                    "Source Control"
                ],
                "summary": "Update source control",
                "description": "Update source control",
                "operationId": "updateSourceControl",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "githuh"
                                    },
                                    "link": {
                                        "type": "examplelink.com",
                                        "example": "1"
                                    },
                                    "id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/source-control/delete/{id}": {
            "get": {
                "tags": [
                    "Source Control"
                ],
                "summary": "Delete source control",
                "description": "Delete source control by id",
                "operationId": "deleteSourceControl",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "{Need to pass id as parameter}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/source-control/list/{id}": {
            "get": {
                "tags": [
                    "Source Control"
                ],
                "summary": "Get single source control",
                "description": "List single source control by id",
                "operationId": "getSourceControlById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "{Need to pass id as parameter}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/team/store": {
            "post": {
                "tags": [
                    "Team"
                ],
                "summary": "Create team",
                "description": "create new team",
                "operationId": "createTeam",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PHP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/team/update": {
            "post": {
                "tags": [
                    "Team"
                ],
                "summary": "Update team",
                "description": "Update team",
                "operationId": "updateTeam",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PHP"
                                    },
                                    "id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/team/delete/{id}": {
            "get": {
                "tags": [
                    "Team"
                ],
                "summary": "Delete team",
                "description": "Delete team by id",
                "operationId": "deleteTeam",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "{Need to pass id as parameter}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "description": "Provide token to access api",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Authorization": []
        }
    ]
}